Optimize delay duration uncaching
Getting rid of false dependencies took a lot of time due to repeated
subsitute() calls with a lot of arguments. In many cases however, there
are only a few parameters/symbols on which delays will depend. We can
therefore get rid of many false dependencies all at once, leaving only a
small amount for the substitute() call.

Furthermore, if the dependencies of a certain delay duration are equal to
the union of all dependencies of all delay durations, we can terminate
early and avoid the (somewhat expensive) ca.symvar() call.